# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSCMR-CET College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  past: "past",
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  category: string;
  venue: string;
  date: string;
  time: string;
  maxParticipants: number;
  registeredCount: number;
  organizer: string;
  contactEmail: string;
  contactPhone: string;
  status: EventStatus;
  imageUrl?: string;
  prizes?: string;
  entryFee?: string;
  createdAt: string;
}

export interface CreateEventInput {
  title: string;
  description: string;
  category: string;
  venue: string;
  date: string;
  time: string;
  maxParticipants: number;
  organizer: string;
  contactEmail: string;
  contactPhone: string;
  imageUrl?: string;
  prizes?: string;
  entryFee?: string;
}

export interface Registration {
  id: number;
  eventId: number;
  studentName: string;
  rollNumber: string;
  branch: string;
  year: string;
  email: string;
  phone: string;
  createdAt: string;
}

export interface CreateRegistrationInput {
  eventId: number;
  studentName: string;
  rollNumber: string;
  branch: string;
  year: string;
  email: string;
  phone: string;
}

export interface Category {
  id: number;
  name: string;
  description: string;
  icon: string;
  color: string;
}

export interface GalleryImage {
  id: number;
  title: string;
  imageUrl: string;
  eventName: string;
  year: string;
}

export interface Stats {
  totalEvents: number;
  upcomingEvents: number;
  totalRegistrations: number;
  totalStudents: number;
}

export type GetEventsParams = {
  category?: string;
  status?: GetEventsStatus;
  search?: string;
};

export type GetEventsStatus =
  (typeof GetEventsStatus)[keyof typeof GetEventsStatus];

export const GetEventsStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  past: "past",
} as const;