# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PSCMR-CET College Event Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Event management
  - name: registrations
    description: Event registrations
  - name: categories
    description: Event categories
  - name: gallery
    description: Event gallery
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get all events
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, past]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /events/{id}/registrations:
    get:
      operationId: getEventRegistrations
      tags: [registrations]
      summary: Get registrations for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"

  /registrations:
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register for an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegistrationInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"

  /registrations/{id}:
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all event categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: Get gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"

  /stats:
    get:
      operationId: getStats
      tags: [events]
      summary: Get event statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        date:
          type: string
        time:
          type: string
        maxParticipants:
          type: integer
        registeredCount:
          type: integer
        organizer:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, past]
        imageUrl:
          type: string
        prizes:
          type: string
        entryFee:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - venue
        - date
        - time
        - maxParticipants
        - registeredCount
        - organizer
        - contactEmail
        - contactPhone
        - status
        - createdAt

    CreateEventInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        date:
          type: string
        time:
          type: string
        maxParticipants:
          type: integer
        organizer:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        imageUrl:
          type: string
        prizes:
          type: string
        entryFee:
          type: string
      required:
        - title
        - description
        - category
        - venue
        - date
        - time
        - maxParticipants
        - organizer
        - contactEmail
        - contactPhone

    Registration:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        studentName:
          type: string
        rollNumber:
          type: string
        branch:
          type: string
        year:
          type: string
        email:
          type: string
        phone:
          type: string
        createdAt:
          type: string
      required:
        - id
        - eventId
        - studentName
        - rollNumber
        - branch
        - year
        - email
        - phone
        - createdAt

    CreateRegistrationInput:
      type: object
      properties:
        eventId:
          type: integer
        studentName:
          type: string
        rollNumber:
          type: string
        branch:
          type: string
        year:
          type: string
        email:
          type: string
        phone:
          type: string
      required:
        - eventId
        - studentName
        - rollNumber
        - branch
        - year
        - email
        - phone

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - description
        - icon
        - color

    GalleryImage:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        eventName:
          type: string
        year:
          type: string
      required:
        - id
        - title
        - imageUrl
        - eventName
        - year

    Stats:
      type: object
      properties:
        totalEvents:
          type: integer
        upcomingEvents:
          type: integer
        totalRegistrations:
          type: integer
        totalStudents:
          type: integer
      required:
        - totalEvents
        - upcomingEvents
        - totalRegistrations
        - totalStudents