# lib/api-zod/src/generated/types/createEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSCMR-CET College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventInput {
  title: string;
  description: string;
  category: string;
  venue: string;
  date: string;
  time: string;
  maxParticipants: number;
  organizer: string;
  contactEmail: string;
  contactPhone: string;
  imageUrl?: string;
  prizes?: string;
  entryFee?: string;
}