# lib/api-zod/src/generated/types/createRegistrationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSCMR-CET College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRegistrationInput {
  eventId: number;
  studentName: string;
  rollNumber: string;
  branch: string;
  year: string;
  email: string;
  phone: string;
}