# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSCMR-CET College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EventStatus } from "./eventStatus";

export interface Event {
  id: number;
  title: string;
  description: string;
  category: string;
  venue: string;
  date: string;
  time: string;
  maxParticipants: number;
  registeredCount: number;
  organizer: string;
  contactEmail: string;
  contactPhone: string;
  status: EventStatus;
  imageUrl?: string;
  prizes?: string;
  entryFee?: string;
  createdAt: string;
}