# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSCMR-CET College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalEvents: number;
  upcomingEvents: number;
  totalRegistrations: number;
  totalStudents: number;
}